Added log function to save test log into log file, and updated test_cards_management.c
