+++ test/src/test.h
@@ -1,0 +1,5 @@
+#ifndef __TEST_HEADER__
+#define __TEST_HEADER__
+
+void write_log(const char* string);
+#endif // __TEST_CARDS_MANAGEMENT_HEADER__
