+++ test/src/test_cards_management.h
@@ -1,0 +1,5 @@
+#ifndef __TEST_CARDS_MANAGEMENT_HEADER__
+#define __TEST_CARDS_MANAGEMENT_HEADER__
+
+int test_cards_mgmt(void);
+#endif // __TEST_CARDS_MANAGEMENT_HEADER__
